CONN-1280: Renamed CORE X12 DS Core per new naming convention, updated copyright date
