--- Product/Production/Services/X12_DocumentSubmissionCore/src/main/java/gov/hhs/fha/nhinc/corex12/docsubmission/realtime/entity/OutboundCORE_X12DocSubmissionOrchestrationContextBuilder.java
@@ -1,89 +1,0 @@
-/*
- * Copyright (c) 2012, United States Government, as represented by the Secretary of Health and Human Services.
- * All rights reserved.
- *
- * Redistribution and use in source and binary forms, with or without
- * modification, are permitted provided that the following conditions are met:
- *     * Redistributions of source code must retain the above
- *       copyright notice, this list of conditions and the following disclaimer.
- *     * Redistributions in binary form must reproduce the above copyright
- *       notice, this list of conditions and the following disclaimer in the documentation
- *       and/or other materials provided with the distribution.
- *     * Neither the name of the United States Government nor the
- *       names of its contributors may be used to endorse or promote products
- *       derived from this software without specific prior written permission.
- *
- * THIS SOFTWARE IS PROVIDED BY THE COPYRIGHT HOLDERS AND CONTRIBUTORS "AS IS" AND
- * ANY EXPRESS OR IMPLIED WARRANTIES, INCLUDING, BUT NOT LIMITED TO, THE IMPLIED
- * WARRANTIES OF MERCHANTABILITY AND FITNESS FOR A PARTICULAR PURPOSE ARE
- * DISCLAIMED. IN NO EVENT SHALL THE UNITED STATES GOVERNMENT BE LIABLE FOR ANY
- * DIRECT, INDIRECT, INCIDENTAL, SPECIAL, EXEMPLARY, OR CONSEQUENTIAL DAMAGES
- * (INCLUDING, BUT NOT LIMITED TO, PROCUREMENT OF SUBSTITUTE GOODS OR SERVICES;
- * LOSS OF USE, DATA, OR PROFITS; OR BUSINESS INTERRUPTION) HOWEVER CAUSED AND
- * ON ANY THEORY OF LIABILITY, WHETHER IN CONTRACT, STRICT LIABILITY, OR TORT
- * (INCLUDING NEGLIGENCE OR OTHERWISE) ARISING IN ANY WAY OUT OF THE USE OF THIS
- * SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
- */
-package gov.hhs.fha.nhinc.corex12.docsubmission.realtime.entity;
-
-import gov.hhs.fha.nhinc.common.nhinccommon.AssertionType;
-import gov.hhs.fha.nhinc.common.nhinccommon.NhinTargetSystemType;
-import gov.hhs.fha.nhinc.orchestration.OutboundDelegate;
-import gov.hhs.fha.nhinc.orchestration.OrchestrationContext;
-import gov.hhs.fha.nhinc.orchestration.OrchestrationContextBuilder;
-import gov.hhs.fha.nhinc.orchestration.OutboundOrchestratable;
-import org.caqh.soap.wsdl.corerule2_2_0.COREEnvelopeRealTimeRequest;
-
-/**
- * @author cmay
- *
- */
-public abstract class OutboundCORE_X12DocSubmissionOrchestrationContextBuilder implements OrchestrationContextBuilder {
-
-    private AssertionType assertionType;
-    private OutboundDelegate nhinDelegate;
-    private COREEnvelopeRealTimeRequest request;
-    private NhinTargetSystemType target;
-
-    @Override
-    public abstract OrchestrationContext build();
-
-    public void init(OutboundOrchestratable message) {
-        setAssertionType(message.getAssertion());
-        setRequest(((OutboundCORE_X12DocSubmissionOrchestratable) message).getRequest());
-        setTarget(((OutboundCORE_X12DocSubmissionOrchestratable) message).getTarget());
-        setNhinDelegate(((OutboundCORE_X12DocSubmissionOrchestratable) message).getNhinDelegate());
-    }
-
-    public AssertionType getAssertionType() {
-        return assertionType;
-    }
-
-    public void setAssertionType(AssertionType assertionType) {
-        this.assertionType = assertionType;
-    }
-
-    public OutboundDelegate getNhinDelegate() {
-        return nhinDelegate;
-    }
-
-    public void setNhinDelegate(OutboundDelegate nhinDelegate) {
-        this.nhinDelegate = nhinDelegate;
-    }
-
-    public COREEnvelopeRealTimeRequest getRequest() {
-        return request;
-    }
-
-    public void setRequest(COREEnvelopeRealTimeRequest request) {
-        this.request = request;
-    }
-
-    public NhinTargetSystemType getTarget() {
-        return target;
-    }
-
-    public void setTarget(NhinTargetSystemType target) {
-        this.target = target;
-    }
-}
